IN-111 refactor: after merge
