+++ src/pages/profile/settings/index.tsx
@@ -1,55 +1,38 @@
-import { useState } from 'react'
+import React from 'react'
 
-import { ImageOutline } from '@/assets/icons/outlineIcons'
-import { Button } from '@/common'
+import { Tabs, TabsType, Typography } from '@/common'
 import { withProtection } from '@/common/HOC/withProtection'
-import { AvatarDialog } from '@/features/profile/settings/avatarDialog'
-import { PhotoResult } from '@/features/profile/settings/avatarDialog/model'
-import Image from 'next/image'
+import { GeneralInfo } from '@/features/profile/settings/generalInfo/ui'
 
-// todo заменить данную страницу на страницу из задачи IN-99/IN-107
+const tabsData: TabsType[] = [
+  {
+    content: <GeneralInfo />,
+    id: 'tab1',
+    title: 'General Information',
+  },
+  {
+    content: <Typography variant={'reg16'}>Devices</Typography>,
+    id: 'tab2',
+    title: 'Devices',
+  },
+  {
+    content: <Typography variant={'reg16'}>Account Management</Typography>,
+    id: 'tab3',
+    title: 'Account Management',
+  },
+  {
+    content: <Typography variant={'reg16'}>My payments</Typography>,
+    id: 'tab4',
+    title: 'My payments',
+  },
+]
 
-const ProfileSettings = () => {
-  const [photoObj, setPhotoObj] = useState<PhotoResult>()
-
+const Settings = () => {
   return (
-    <div
-      className={
-        'flex justify-center items-center flex-col w-full h-[500px] gap-6'
-      }
-    >
-      <div
-        className={
-          'w-[192px] aspect-square flex justify-center items-center rounded-full bg-dark-500'
-        }
-      >
-        {photoObj?.photo ? (
-          <Image
-            alt={'avatar'}
-            className={'object-cover'}
-            height={200}
-            src={photoObj.photo}
-            width={200}
-          />
-        ) : (
-          <div
-            className={
-              'w-[192px] aspect-square flex justify-center items-center rounded-full bg-dark-500'
-            }
-          >
-            <ImageOutline
-              height={48}
-              width={48}
-            />
-          </div>
-        )}
-      </div>
-      <AvatarDialog
-        onModalPhotoSave={setPhotoObj}
-        triggerButton={<Button variant={'outline'}>Add profile photo</Button>}
-      />
+    <div className={'mt-[42px]'}>
+      <Tabs tabsData={tabsData} />
     </div>
   )
 }
 
-export default withProtection(ProfileSettings, true)
+export default withProtection(Settings, false)
